Update Attributions.js
Same apply in all of these cases i guees. So can we separate things? HTML or view related things will be in html file and Controller things will be in js file .. Also sometimes it will be necessary to keep js reference in html file but that's okay. We should keep things separated(clean code)